feat: update error view style
