+++ app/components/Views/ErrorBoundary/index.js
@@ -4,92 +4,15 @@ import PropTypes from 'prop-types';
 import RevealPrivateCredential from '../RevealPrivateCredential';
 import Logger from '../../../util/Logger';
 import routes from '../../../common/routes';
-import { colors, fontStyles } from '../../../styles/common';
 import { ScrollView } from 'react-native-gesture-handler';
 import Clipboard from '@react-native-community/clipboard';
 import { strings } from '../../../../locales/i18n';
 import Icon from 'react-native-vector-icons/FontAwesome';
+import styles from './styles';
+import drawables from '../../../common/drawables';
 
 // eslint-disable-next-line import/no-commonjs
-const metamaskErrorImage = require('../../../images/metamask-error.png');
-
-const styles = StyleSheet.create({
-	container: {
-		flex: 1
-	},
-	content: {
-		paddingHorizontal: 24,
-		flex: 1
-	},
-	header: {
-		alignItems: 'center'
-	},
-	errorImage: {
-		width: 50,
-		height: 50,
-		marginTop: 24
-	},
-	title: {
-		color: colors.black,
-		fontSize: 24,
-		lineHeight: 34,
-		...fontStyles.bold
-	},
-	subtitle: {
-		fontSize: 14,
-		lineHeight: 20,
-		color: colors.grey500,
-		marginTop: 8,
-		textAlign: 'center',
-		...fontStyles.normal
-	},
-	errorContainer: {
-		backgroundColor: colors.red000,
-		borderRadius: 8,
-		marginTop: 24
-	},
-	error: {
-		color: colors.black,
-		padding: 8,
-		fontSize: 14,
-		lineHeight: 20,
-		...fontStyles.normal
-	},
-	button: {
-		marginTop: 24,
-		borderColor: colors.blue,
-		borderWidth: 1,
-		borderRadius: 50,
-		padding: 12,
-		paddingHorizontal: 34
-	},
-	buttonText: {
-		color: colors.blue,
-		textAlign: 'center',
-		...fontStyles.normal,
-		fontWeight: '500'
-	},
-	textContainer: {
-		marginTop: 24
-	},
-	text: {
-		color: colors.black,
-		fontSize: 14,
-		lineHeight: 20,
-		...fontStyles.normal
-	},
-	link: {
-		color: colors.blue
-	},
-	reportTextContainer: {
-		paddingLeft: 14,
-		marginTop: 16,
-		marginBottom: 24
-	},
-	reportStep: {
-		marginTop: 14
-	}
-});
+const metamaskErrorImage = drawables.logo;
 
 const Fallback = props => (
 	<SafeAreaView style={styles.container}>
